+++ hack/checks.sh
@@ -1,13 +1,13 @@
 #!/bin/bash
-cd $(dirname $0)/..
+cd $(dirname $0)/../
 
 set -xeuo pipefail
 go mod tidy
 go fmt ./...
 go vet ./...
-go test ./... -v --failfast
 
 bash hack/go-licenses.sh
+bash hack/go-test.sh
 bash hack/gocyclo.sh
 bash hack/golangci-lint.sh
 bash hack/misspell.sh
