+++ pkg/handler/router.go
@@ -2,15 +2,16 @@ package handler
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/zetaoss/runbox/pkg/handler/lang"
 	"github.com/zetaoss/runbox/pkg/handler/notebook"
+	"github.com/zetaoss/runbox/pkg/handler/run"
+	"github.com/zetaoss/runbox/pkg/handler/status"
 )
 
 func NewRouter() *gin.Engine {
 	gin.SetMode(gin.ReleaseMode)
 	r := gin.Default()
-	r.GET("/-/healthy", healthy)
-	r.POST("/lang", lang.Run)
-	r.POST("/notebook", notebook.Run)
+	r.GET("/-/healthy", status.Healthy)
+	r.POST("/api/notebook", notebook.Run)
+	r.POST("/api/run", run.Run)
 	return r
 }
