+++ pkg/handler/run/errors.go
@@ -1,4 +1,4 @@
-package lang
+package run
 
 type Error string
