+++ pkg/runner/run/run.go
@@ -1,4 +1,4 @@
-package lang
+package run
 
 import (
 	"fmt"
