+++ pkg/runner/run/run_test.go
@@ -1,4 +1,4 @@
-package lang
+package run
 
 import (
 	"testing"
